Attached jQuery UI datepicker to date input fields, corrected IDs data types of entities(from long to int) and made pretty presentation of tables elements on admin pages

+++ src/main/java/ru/ildar/controllers/admin/FacultiesController.java
@@ -48,7 +48,7 @@ public class FacultiesController
     }
 
     @RequestMapping(value = "add", method = RequestMethod.GET)
-    public ModelAndView addNewFaculty(@RequestParam("unId") long uniId, ModelMap model)
+    public ModelAndView addNewFaculty(@RequestParam("unId") int uniId, ModelMap model)
     {
         University un = universityService.getById(uniId);
         model.addAttribute("university", un);
